SDK: Fix MSVC 2022 warning
**Description:** Fixed warning with a static_cast.

**Testing Performed:** Build on MSVC 2022

+++ SDK/simCore/Time/TimeClass.cpp
@@ -476,7 +476,7 @@ int TimeStamp::strptime(const std::string& timeStr, const std::string& format, s
         *remainder = timeStr;
       return 1;
     }
-    setTime(1970, asTime);
+    setTime(1970, static_cast<double>(asTime));
     return 0;
   }
   catch (const simCore::TimeException&)
